AI实现/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
numpy==1.26.4
sentence-transformers==2.5.1
faiss-cpu==1.8.0
requests==2.31.0
openpyxl==3.1.2